+++ modules/polyfills/src/node/images/parse-image.node.ts
@@ -1,4 +1,9 @@
+// loaders.gl, MIT license
+
 import getPixels from 'get-pixels';
+
+/** Declares which image format mime types this loader polyfill supports */
+export const NODE_FORMAT_SUPPORT = ['image/png', 'image/jpeg', 'image/gif'];
 
 // Note: These types are also defined in @loaders.gl/images and need to be kept in sync
 type NDArray = {
@@ -16,9 +21,7 @@ export async function parseImageNode(arrayBuffer: ArrayBuffer, mimeType: string)
   }
 
   const buffer = arrayBuffer instanceof Buffer ? arrayBuffer : Buffer.from(arrayBuffer);
-
   const ndarray = await getPixelsAsync(buffer, mimeType);
-
   return ndarray;
 }